+++ src/BaseTransformer.php
@@ -3,6 +3,7 @@
 namespace Saritasa\Transformers;
 
 use Carbon\Carbon;
+use DateTime;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Database\Eloquent\Model;
 use League\Fractal\TransformerAbstract;
@@ -14,9 +15,17 @@ use League\Fractal\TransformerAbstract;
 class BaseTransformer extends TransformerAbstract implements IDataTransformer
 {
     /**
+     * Default format of date that will be used to convert date fields
+     *
+     * @var string
+     */
+    protected $defaultDateTimeFormat = Carbon::ISO8601;
+
+    /**
      * Transform model into array
      *
      * @param Arrayable $model Model to be transformed
+     *
      * @return array
      */
     public function transform(Arrayable $model)
@@ -25,15 +34,19 @@ class BaseTransformer extends TransformerAbstract implements IDataTransformer
         if ($model instanceof Model) {
             $result = $this->datesToISO8601($result, $model);
             $result = $this->removeHiddenFields($result, $model);
+        } else {
+            $result = $this->formatDatesInArray($result);
         }
+
         return $result;
     }
 
     /**
-     * Convert dates to ISO 8601 format (includes time zone)
+     * Converts dates in model to ISO 8601 format (includes time zone)
      *
      * @param array $result model, serialized to array
      * @param Model $model original Eloquent model
+     *
      * @return array
      */
     protected function datesToISO8601(array $result, Model $model)
@@ -42,9 +55,10 @@ class BaseTransformer extends TransformerAbstract implements IDataTransformer
         if ($dateFields && count($dateFields) > 0) {
             foreach ($dateFields as $field) {
                 $value = $model->getAttributeValue($field);
-                $result[$field] = $value ? $value->format(Carbon::ISO8601) : $value;
+                $result[$field] = $value ? $value->format($this->defaultDateTimeFormat) : $value;
             }
         }
+
         return $result;
     }
 
@@ -54,6 +68,7 @@ class BaseTransformer extends TransformerAbstract implements IDataTransformer
      *
      * @param array $result model, serialized to array
      * @param Model $model original Eloquent model
+     *
      * @return array
      */
     protected function removeHiddenFields(array $result, Model $model)
@@ -66,6 +81,35 @@ class BaseTransformer extends TransformerAbstract implements IDataTransformer
                 }
             }
         }
+
+        // If Eloquent model has both $hidden and $visible arrays then need to satisfy both restrictions
+        $visibleFields = $model->getVisible();
+        if ($visibleFields && count($visibleFields) > 0) {
+            foreach ($result as $field => $value) {
+                if (!in_array($field, $visibleFields)) {
+                    unset($result[$field]);
+                }
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Converts dates in array to ISO 8601 format (includes time zone)
+     *
+     * @param array $result Array in which need to convert dates
+     *
+     * @return array Result array with converted dates
+     */
+    protected function formatDatesInArray(array $result): array
+    {
+        foreach ($result as $attribute => $value) {
+            if ($value instanceof DateTime) {
+                $result[$attribute] = $value->format($this->defaultDateTimeFormat);
+            }
+        }
+
         return $result;
     }
 }